changed av_velocity to al floats #

+++ d2q9-bgk.c
@@ -370,7 +370,7 @@ int collision(const t_param params, t_speed* cells, t_speed* tmp_cells, int* obs
 double av_velocity(const t_param params, t_speed* cells, int* obstacles)
 {
   int    tot_cells = 0;  /* no. of cells used in calculation */
-  double tot_u;          /* accumulated magnitudes of velocity for each cell */
+  float tot_u;          /* accumulated magnitudes of velocity for each cell */
 
   /* initialise */
   tot_u = 0.0;
@@ -397,7 +397,7 @@ double av_velocity(const t_param params, t_speed* cells, int* obstacles)
 
 
         /* x-component of velocity */
-        double u_x = (cells[cellAccess].speeds[1]
+        float u_x = (cells[cellAccess].speeds[1]
                       + cells[cellAccess].speeds[5]
                       + cells[cellAccess].speeds[8]
                       - (cells[cellAccess].speeds[3]
@@ -405,7 +405,7 @@ double av_velocity(const t_param params, t_speed* cells, int* obstacles)
                          + cells[cellAccess].speeds[7]))
                      / local_density;
         /* compute y velocity component */
-        double u_y = (cells[cellAccess].speeds[2]
+        float u_y = (cells[cellAccess].speeds[2]
                       + cells[cellAccess].speeds[5]
                       + cells[cellAccess].speeds[6]
                       - (cells[cellAccess].speeds[4]
